Rename develop branch to development and add Code Climate and Cypress (#1)
* Rename develop branch to development

We should try and be as clear as possible. In this case, the word
"development" is more formal than "develop" and is used as the main
branch of this repo.

* Set up Cypress and Code Climate

Since we are now open-sourcing this library. We have swapped out a new
Code Climate and Cypress project ID to take advantage of their free
plans.

Co-authored-by: Nicholas Shook <nick@neonlaw.com>