Fix Staging environment
On staging, we are experiencing this error in Nginx:

```
crypt_r() failed (22: Invalid argument)
```

This error is likely because crypt_r, a C utility used by Nginx, cannot
parse our htpasswd file for HTTP Basic authentication. This PR addresses
that by changing how we take the GitHub Actions environment variable and
adds it to the workflow, which then injects the Docker container.